handler/requirements.txt
kafka-python==2.0.1
pymongo==3.10.1
